Added new definitions, minor fixes
