--- openquake/hazardlib/tests/gsim/data/PP2004/PP2004_MEAN.csv
@@ -1,13 +1,0 @@
-rup_mag,dist_rjb,site_vs30,result_type,damping,PGA
-5,0,1000,MEAN,5,0.126
-5,0,500,MEAN,5,0.189
-5,70,1000,MEAN,5,0.0116
-5,70,500,MEAN,5,0.0173
-6,0,1000,MEAN,5,0.214
-6,0,500,MEAN,5,0.320
-6,70,1000,MEAN,5,0.0197
-6,70,500,MEAN,5,0.0293
-7,0,1000,MEAN,5,0.363
-7,0,500,MEAN,5,0.542
-7,70,1000,MEAN,5,0.0333
-7,70,500,MEAN,5,0.0497
